Don't obfuscate database names
At the start of the project i thought if would be cool to have db names
obfuscated but it's a nonsense. It only adds more logic and brain fucks
but no benefit.

After this commit manuall database renaming has to be done othrwise
SupTasks will create new empty database for existing users.