Remove time records form from tasks index
Removed because i don't think it's key part on this page. I want to have
a fast overview on this page about my tasks. I want to track my time but
as most easily as possible. And all spend time review should be done on
other page build specially for this purpose.